chore(tests): move db trait to concerns
